Move the login module - as a package - into app
