+++ app/src/main/kotlin/uk/gov/govuk/login/ui/ErrorScreen.kt
@@ -6,10 +6,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
+import uk.gov.govuk.R
 import uk.gov.govuk.design.ui.component.ErrorPage
 import uk.gov.govuk.design.ui.theme.GovUkTheme
 import uk.gov.govuk.login.ErrorViewModel
-import uk.gov.govuk.login.R
 
 @Composable
 internal fun ErrorRoute(
